pyproject.toml
[tool.poetry]
name = "music-tool-bot"
version = "2.6.0"
description = "Music Tool Bot is the ultimate bot for managing your music files effortlessly! 🎵✨"
readme = "README.md"
authors = ["Amir Hosein Salimi <iamirhoseinsalimi@gmail.com>"]
license = "LICENSE"
keywords = ["python", "music", "telegram-bot", "python-telegram-bot", "tag-editor"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10.12"
music-tag = ">=0.4.3,<0.5.0"
python-dotenv = ">=1.0.1,<2.0.0"
psutil = ">=6.1.1,<7.0.0"
persiantools = ">=4.2.0,<5.0.0"
masonite-orm = ">=2.22.3,<3.0.0"
psycopg2-binary = ">=2.9.9,<3.0.0"
cryptography = ">=43.0.3,<44.0.0"
httpcore = "*"
pillow = "*"
python-telegram-bot = { version = ">=22.0,<23.0", extras = ["job-queue"] }
pyi18n-v2 = "^1.2.2"

[tool.poetry.group.dev.dependencies]
pylint = ">=3.3.6,<4.0.0"
autopep8 = ">=2.3.1,<3.0.0"
jurigged = ">=0.6.0,<0.7.0"
pydevd-pycharm = "241.17890.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
